backend/requirements.txt
Django>=5.0
djangorestframework
django-cors-headers
Pillow
psycopg2-binary